stock-agent-v6/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.0
python-dotenv>=1.0
langsmith>=0.1.0
pytest>=7.0
akshare>=1.12
pandas>=2.0
sqlalchemy>=2.0
alembic>=1.13
apscheduler>=3.10
fastapi>=0.100
uvicorn[standard]>=0.30
sse-starlette>=2.0
